refactor(matrices): move copyin_list, copyin_matrix to RepMatrix
